Added test cases for excludes, new testhelper.py to collect test-related functions
