+++ tests/test_find_backups.py
@@ -1,33 +1,28 @@
-from datetime import datetime
+from os import path
 
 from causync import CauSync
 import config
 
-dirnames = ['20040101', '20050101', '20060101', '20070101', '20080101',
-            '20090101', '20100101', '20110101', '20120101', '20130101',
-            '20140101', '20150101', '20160101', '20160201', '20160411',
-            '20170101', '20170701', '20170801', '20170901', '20171001',
-            '20171101', '20171201', '20180101', '20180201', '20180301',
-            '20180401', '20180402', '20180403', '20180404', '20180405',
-            '20180406', '20180407', '20180408', '20180409', '20180410',
-            '20180411']
+from tests.testhelper import *
+
+src = path.realpath("/tmp/causync_src")
+dst = path.realpath("/tmp/causync_dest")
 
 
 def test_find_latest_backups():
-    cs = CauSync(config, "/tmp/causync_src", "/tmp/causync_dest", task='sync')
+    cs = CauSync(config, src, dst, task='sync')
     cs.config.DATE_FORMAT = "%Y%m%d"
 
-    dirnames_result = ['/tmp/causync_dest/20180411',
-                       '/tmp/causync_dest/20180410',
-                       '/tmp/causync_dest/20180409',
-                       '/tmp/causync_dest/20180408',
-                       '/tmp/causync_dest/20180407']
+    paths = ['20180411', '20180410', '20180409', '20180408', '20180407']
 
+    dirnames_result = [path.realpath("{}/{}".format(dst, i)) for i in paths]
+
+    print(cs.find_latest_backups(dirnames))
     assert cs.find_latest_backups(dirnames) == dirnames_result
 
 
 def test_find_old_backups_daily():
-    cs = CauSync(config, "/tmp/causync_src", "/tmp/causync_dest", task='cleanup')
+    cs = CauSync(config, src, dst, task='cleanup')
     cs.config.DATE_FORMAT = "%Y%m%d"
     cs.config.BACKUPS_TO_KEEP = {'yearly': 10, 'monthly': 6,
                                  'weekly': 4, 'daily': 7}
@@ -59,7 +54,7 @@ def test_find_old_backups_daily():
 
 
 def test_find_old_backups_weekly():
-    cs = CauSync(config, "/tmp/causync_src", "/tmp/causync_dest", task='cleanup')
+    cs = CauSync(config, src, dst, task='cleanup')
     cs.config.DATE_FORMAT = "%Y%m%d"
     cs.config.BACKUPS_TO_KEEP = {'yearly': 10, 'monthly': 6,
                                  'weekly': 4, 'daily': 7}
@@ -76,7 +71,7 @@ def test_find_old_backups_weekly():
 
 
 def test_find_old_backups_monthly():
-    cs = CauSync(config, "/tmp/causync_src", "/tmp/causync_dest", task='cleanup')
+    cs = CauSync(config, src, dst, task='cleanup')
     cs.config.DATE_FORMAT = "%Y%m%d"
     cs.config.BACKUPS_TO_KEEP = {'yearly': 10, 'monthly': 6,
                                  'weekly': 4, 'daily': 7}
@@ -103,7 +98,7 @@ def test_find_old_backups_monthly():
 
 
 def test_find_old_backups_yearly():
-    cs = CauSync(config, "/tmp/causync_src", "/tmp/causync_dest", task='cleanup')
+    cs = CauSync(config, src, dst, task='cleanup')
     cs.config.DATE_FORMAT = "%Y%m%d"
     cs.config.BACKUPS_TO_KEEP = {'yearly': 10, 'monthly': 6,
                                  'weekly': 4, 'daily': 7}
